# HEIGVD-AMT-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
  /badges:
    get:
      description: 
        The Badges endpoint returns information about badges belonging to the application whose id is present in the authentication token. The response includes the id, the name, the description, and the image URL of each badge. This response is a list of badge sorted by the id.
      produces:
        - application/json
      tags:
        - Badges
      parameters:
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. This token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK, an array of badges.
          schema:
            type: array
            items:
              $ref: '#/definitions/BadgeOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired.
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.  
        500:
          description: Internal server error, if database isn't reachable.
              
    post:
      description: Create a new badge belonging to the application whose id is present in the authentication token.
      consumes:
        - application/json
      parameters:
        - in: body
          name: badge
          description: Badge object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/BadgeInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        201:
          description: Created, returns the URL to access the new badge.
          schema:
            $ref: '#/definitions/LocationBadge'
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.  
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if badge name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.
  /badges/{id}:
    get:
      description: Returns the badge which has the specified id belonging to the application whose id is present in the authentication token.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: The id of the desired badge.
          required: true
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/BadgeOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.  
        500:
          description: Internal server error, if database isn't reachable.
          
    put:
      description: Modify completely the existing badge which has the specified id belonging to the application whose id is present in the authentication token.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: The id of the desired badge.
          required: true
        - name: badge
          in: body
          description: Badge to be updated
          required: true
          schema:
              $ref: '#/definitions/BadgeInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.  
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if badge name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.

    delete:
      description: Delete the badge which has the specified id belonging to the application whose id is present in the authentication token.
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: The id of the desired badge.
          required: true
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.  
        500:
          description: Internal server error, if database isn't reachable.
...
definitions:
  BadgeOutputDTO:
    type: object
    required: [badge_id, name, description, imageURL]
    properties:
      badge_id:
        type: integer
        format: int64
        description: Unique identifier representing a specific badge.
      name:
        type: string
        description: Display name of badge.
      description:
        type: string
        description: Description of badge.
      imageURL:
        type: string
        description: Image URL representing the badge.
  BadgeInputDTO:
    type: object
    required: [name,description,imageURL]
    properties:
      name:
        type: string
        description: Display name of badge.
      description:
        type: string
        description: Description of badge.
      imageURL:
        type: string
        description: The URL of the image          
  LocationBadge:
      type: string
      required: [locationBadge]
      properties:
        locationBadge:
          type: string
          description: Unique URL to access the new object created.
...